refactor(coil):  helper functions + update identities
